refactor(native-chat): isolate resolved bridge view
